model-server/requirements.txt
flask
transformers==4.40.0
torch
Pillow
numpy
faiss-cpu
requests
python-dotenv
gunicorn
